feat(hosts): Add a quick tool to format hosts file
This is a quick and dirty tool to format hosts file.
It parse all blocked domains, sort them according TLD then domain and
print them by domain.TLD section (with a header before each section).

This is mainly a proof of concept and the script could be improved in a
later time to take CLI arguments, add AdAway hosts file header, etc…