remove uncommon hosts pt 2
as with previous commit, this time I compared the Adaway list to Cisco/OpenDNS' Umbrella 1 mllion list for 4 dates in this year.  (1/15/2020, 7/1/2020, 8/1/2020, 8/19/2020).  I just wanted to trim it up with the premise that any top 1 million DNS lookups from last year may be worth less as companies shut down and/or servers/domain names may no longer be active.

If I notice a few that are still active in mobile apps I'll add them again.  Just trying to clean out the cruft.