remove entries based on DNS lookup inactivity
step 1 - http://s3-us-west-1.amazonaws.com/umbrella-static/index.html

Downloaded the daily logs from August 1st, 2020 to October 13th, 2020 top 1 million DNS lookups for all of Cisco's OpenDNS/Umbrella users.

If an entry showed on any one of those days I kept it.  Otherwise it was scrapped.  The premise being if a DNS lookup is not popular enough to show on a massive DNS log file of that many user devices it is not worth trying to block it.