remove old stuff
based on Cisco Umbrella Top 1million data
downloaded about 2 weeks of old hits.  Anything that didn't have at least 1 hit on the top 1 million on any one day I removed.  And then looked back over to see if any of them had one entry for that domain left and looked a few of them up.  Looks like it caught a few companies that went out of business but their home page remained- so it was probably worth getting rid of the fluff.